+++ tests/test_project/app_add_not_null_column_followed_by_default/migrations/0002_a_not_null_field.py
@@ -3,7 +3,6 @@
 from __future__ import unicode_literals
 
 from django.db import migrations, models
-
 from django_add_default_value import AddDefaultValue