+++ tests/test_project/app_drop_table/models.py
@@ -3,6 +3,5 @@ from __future__ import unicode_literals
 
 from django.db import models
 
-
 # class A(models.Model):
 #    field_a = models.IntegerField()