+++ tests/test_project/urls.py
@@ -13,7 +13,7 @@ Including another URLconf
     1. Import the include() function: from django.urls import include, path
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
-from django.contrib import admin
 from django.conf.urls import url
+from django.contrib import admin
 
 urlpatterns = [url("admin/", admin.site.urls)]