+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add Python 3.9 support
 * Make data migration model import error less strict (issue #121)
 * Add warning detection on RunPython call when model variable name is not the same as model class name
+* Run checks on RunSQL migration operations
 
 Switched from Travis CI to GitHub Actions for tests